convert commands to cogs
